Optimizing gzip parameters
The flags -n, -c and -9 were added to the invocation of gzip. With
the statistics directory, this led to an apparent reduction of 1 byte
over the default settings. With the ASCII encoding the effect will
be increased. Also the size of information stored is likely to increase,
which further motivates to become more ambitious in compression.

The meaning of the parameters is

 -9 (or --best): best possible encryption
 -n            : do not store name etc of filenames, also not any temporary ones
 -c            : use the stdio

It is not unlikely that -n (likely) and -c (very likely) are both
implicitly set. Anyway.