Altered the domain_code-provider in order to support domaincode-files.
Given a dcf-path, the provider tries to return the domaincode related
to a batadv_dev, like trier did earlier in their code.

If it cannot find a match, it returns whatever was provided as
'domaincode' via '-n' in the commandline, which makes it effectively a
default value for unkown domains and a nice fallback.